Adapt frontend to new api
