+++ game-frontend/src/Game.tsx
@@ -4,24 +4,25 @@ import Position from '../../darwin-types/Position';
 import CanvasWrapper from './components/visual/canvas-wrapper';
 import useWebsocketData from './service/useWebsocketData';
 import Unit from './components/canvas-objects/Unit';
+import { State } from '../../darwin-types/State';
 
 interface Unit {
   position: Position;
 }
 
 function Game(): JSX.Element {
-  const unit: Unit | null = useWebsocketData();
+  const gameState: State = useWebsocketData();
 
-  if (!unit) {
+  if (!gameState) {
     return <p>Loading</p>;
   }
-
-  const { x, y } = (unit as Unit).position;
 
   return (
     <CanvasWrapper>
       <Stage width={200} height={200}>
-        <Unit position={{ x: x * 10, y: y * 10 }} />
+        {gameState.objectIds.map(objectId => (
+          <Unit {...gameState.objectMap[objectId]} key={objectId} />
+        ))}
       </Stage>
     </CanvasWrapper>
   );
